refactor: seperate input component as general
